requirements.txt
ollama
textual
pydantic_settings
gptfunctionutil